backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiosqlite==0.20.0
groq==0.9.0
deepgram-sdk==3.2.7
elevenlabs==1.2.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
pyyaml==6.0.1
httpx==0.27.0
websockets==12.0
pytest==8.2.0
pytest-asyncio==0.23.6
